[IMP] account_accountant: fix regex for payment matching
In this commit: https://github.com/odoo/enterprise/commit/c0f8c5d17b9d78b640d744d8bfb6a2bb4002776a, we changed the way payment are matched with statement lines.

Before, we were splitting the aml.ref based on ' ', then check for each word if there is an equivalent in the st_line label.

But this is not restrictive enough. So now, there is the rules we apply, in the right order :
1 - Check if there is a payment ref like SO|INV|BILL|...xxxx/xxx-yy in the aml.ref
2 - Check for full match with aml.ref
3 - Check if there is a word more than 16 characters long who match with aml.ref
4 - Check for full match with aml.move_name
5 - Check for full match with move.payment_reference

If we have any multiple match, we don't reconcile.

no-task

X-original-commit: 584a3b417d10042af20763ac39b932b61fbfa2df
Part-of: odoo/enterprise#88085
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>