[IMP] delivery_ups: remove UPS account from address details
The 'UPS Account Number' field was always visible on the portal user's address form, regardless of context.
Introduced 5 years ago with 10892c12c62fb7d82ed55f53ffe3815298b89018, it was recently moved from `website_sale_ups` to `delivery_ups`, as it was used in the core delivery logic, not only the ecommerce one.

Nevertheless, it has been considered unnecessary to keep it on the portal form, as it is displayed and chosen when necessary in the e-commerce checkout flow).  Furthermore, this field was shown regardless of whether the ups carrier was enabled and published (or was ever used by the current customer).

We remove it for now to simplify the default address form (and runbot form). If needed, we'll consider re-introducing it in the future (but with additional restrictions).

task-4669485

closes odoo/enterprise#87200

Related: odoo/odoo#204726
Related: odoo/upgrade#7826
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>