[FIX] account_budget: Remove base.main_company from a test
This fix addresses a test failure in runbot. The test previously relied on base.main_company, which causes issues when running the test suite on master without the --with-demo flag enabled.

To ensure compatibility, the test has been updated to use self.company directly instead of base.main_company

runbot-226645

closes odoo/enterprise#87824

Signed-off-by: William André (wan) <wan@odoo.com>