[FIX] account_accountant: None for short common substring of 2 labels
The _get_common_substring method ought to return None if the
substring is shorter than 10. The check is only done while looping through
the normalised labels, but the loop is skipped if we have exactly 2 labels,
thus returning the substring irrespective of its length.

Instead, we should do the actual length check (and substitute with None
as needed) right before returning.

The phrasing of the method's docstring is also updated to clarify that
in the case of identical normalised labels, the _normalised_ label is
returned rather than its original form.

closes odoo/enterprise#88094

X-original-commit: ffba6802726985c4825cd3710b60708ec5402377
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>