[FIX] l10n_ke_edi_oscu: Restore details of timed out invoices
When a timeout occurs during invoice submission to eTIMS, we had no way
to know if the invoice was actually received. Retrying could lead to
duplicate invoices being accepted by the government.

With the release of the selectInvoiceDetails endpoint, we now check if the invoice was already submitted before retrying. If found, we treat it as successfully sent and retrieve its details (QR code, etc). If not, we resend.

task-4863787

closes odoo/enterprise#88957

X-original-commit: 13b201172b67dc38042cfd0229137d4b2baf412e
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Signed-off-by: Seif Gneedy (segn) <segn@odoo.com>