[FIX] hr_gantt: add correct model for employee popover
When grouping by employee in a gantt view, you can click on the employee's avatar to open a popover.
The resModel of that popover card must be correctly set in order to fetch the data on the right model (employee or employee public).
This commit fixes the issue by specifying the record model.

task-4873046

closes odoo/enterprise#87834

Related: odoo/odoo#214517
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>