[FIX] l10n_hk_hr_payroll: give payroll administrator access rights
Since the access rights to see the hk rental is given to the payroll
officer and the rights to modify the status is given to the employee
administrator, you have payroll adminsitrator that cannot change the
status and employee administrator that cannot access the model.

As the model seems to be dedicated to Payroll and you need to be payroll
officer to techniaclly change the status, we modify the view to allow
this operation to payroll Adminsitrator.

closes odoo/enterprise#88084

X-original-commit: 5409decf105d7f507c9c019f14874d398f162597
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>