[ADD] l10n_cz_intrastat: Intrastat csv for Czech Republic
We are adding intrastat report export format in csv for the Czech
Republic we can be downloaded when submitting the tax return for the
intrastat.

Since the tax return wasn't present, we are also adding it.

We are not checking if we dont need to generate this file as we are
not auto-installing the module and consider that It is enough for now.

task-4826249

closes odoo/enterprise#85470

Related: odoo/odoo#222598
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>