[FIX] account_budget: fix computation of is_above_budget
Steps to reproduce:
- Install account_budget
- Activate "Budget Management" in Accounting settings

- Go to "Accounting / Accounting / Analytic Budget"
- Create a new budget:
  * Budget Type: Expense
  * Period: [this month]
  * Budget Lines: [a line with an analytic account (AA) in 2 plans]
- Open the budget

Issue 1:
Create a PO at the current date with a line having only one of the
AA as analytic distribution and an amount higher than the budgeted
amount.
The PO line and the Budget button appear in red, but it shouldn't
because the analytic distribution of the line doesn't match all the
accounts defined on the budget line.
The line will not impact the budget, but it appears in red, which is
confusing.

Cause 1:
When computing the budget lines linked to the PO line, the domain
that is used is only taking into account the plans that are set on
the PO line, ignoring the plans that are not set.
For the Budget button, all the PO lines are used instead of just
those having the correct analytic distribution.

Issue 2:
- Create a PO at the current date with 2 lines:
1) a line with the same AA than the budget line and an amount higher
than the budget amount
2) a line with the same AA than the budget line and an amount lower
than the budget amount
- Save the PO
The first line should be red, but it is not.

Cause 2:
When computing the budget lines linked to the PO lines, only the last
PO line has its value set.

opw-4757067

closes odoo/enterprise#88192

X-original-commit: fbcb4f42d9745e849c55fb7348cb2205b164f1c4
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>