[FIX] l10n_cl_edi_pos: display correct settle invoice price in Chile
**Problem:**
When trying to settle an order in POS with a Chilean company,
the amount to be settle will be 1$, regardless of how much
needed to be settled.

**Steps to reproduce:**
- Set the company to a Chilean company.
- Make an invoice for a Chilean customer and confirm it.
- Go to pos, and click settle invoices for your client, chose the invoice you just made.
- The amount will be 1$

**Why the fix:**
The field *pos_amount_unsettled* was not loaded into the invoice if the company
is from Chile. It resulted in it being undefined, thus it was set to 1 afterwards.
We now load this variable into the invoice when loading the data.

opw-4856764

closes odoo/enterprise#89003

X-original-commit: 20ea0423c001ac1aee9aff69b1f3165d78ba6015
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Arthur Nanson (artn) <artn@odoo.com>