[IMP] {sale_renting_}planning: bind slot and SO
In the case of Rental of services, such as bike rental, car rental, hotel room rental...
Relying on the resources makes a lot of sense. It is already enabled to book a resource in such case!

However, when no resources are available for the role specified by the product, a booking (Rental Order)
can anyway be confirmed which therefore leads to conflicts in planning since no resources are available for this order!
In the case of an hotel, they can't make a new room!

Such use case can be worked around in Rental with some automations.
However, eCommerce needs to be fed with availabilities in order to prevent double booking from website.

task-4863072

closes odoo/enterprise#88174

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>