[FIX] voip: prevent error in voip config if no phonecall activity type exists
Currently, an error occurs when the system initializes the voip configuration
for the currently logged-in internal user.

Steps to Reproduce:

 - Install the voip module.
 - Delete the activity type where the Action is 'Phonecall', and then
   refresh the page.

IndexError: tuple index out of range.

This error occurs when the system initializes the voip configuration and no
phonecall activity type exists. The system attempts to search for activity
types with the 'phonecall' category, which returns an empty recordset.
Then it tries to access the first record of this empty recordset[1], which
raises an error.

[1] https://github.com/odoo/enterprise/blob/2896974e9eda61bfe5e1d67423dc56f6ab17b2af/voip/models/res_users.py#L128

This commit ensures that if activity type with the 'phonecall' category
is present, it store the activity type id, otherwise, it stores False
as the activity type id.

sentry-6689050392

closes odoo/enterprise#88876

X-original-commit: 4c4253c0ec3d1e8f11677fe18d73f0d015f5569e
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>
Signed-off-by: Ashutosh Sharma (assh) <assh@odoo.com>