[FIX] website_studio: add "selectModel" as a shared function
Since [1] when actions were converted to classes, the `selectModel`
method is called through the dependencies/shared function mechanism.
Unfortunately, the function could not be found because it was not
registered as a shared function.

This commit adds the function to the list of shared functions.

Steps to reproduce:
- Install website_studio
- Drop a form
- Pick the "Mode models" action

=> There was an error because `selectModel` could not be found.

[1]: https://github.com/odoo/enterprise/commit/1f8b599a23330b03bffd898d85fe74a0f0eeda1f

task-4367641

closes odoo/enterprise#88028

Related: odoo/odoo#214959
Signed-off-by: Francois Georis (fge) <fge@odoo.com>