[FIX] social_youtube: make 'youtube_video_category_id' non-readonly
If not specified, the field 'youtube_video_category_id' was treated as
readonly by default. This commit explicitly sets readonly: false,
ensuring that the field is editable. This prevents the value from being
lost during posting, as it will now be correctly written to the database.

Part of PR: #83571
opw-4646432

closes odoo/enterprise#88556

X-original-commit: 9a7dafef9791af478b0e24e81ca17bca605bcc67
Signed-off-by: Maruan Aguerdouh Mohtar (magm) <magm@odoo.com>