[FIX] pos_urban_piper: close session with duplicate method
Currently if setting a urban paiper payment method on a shop and using
it at checkout will not allow you to close the session.

Steps to reproduce:
-------------------
* Go in the setting and set up urban piper (no need for credentials) testing
with random keys will do
* Go into the payment methods and take one of th eurban piper one and
assign it to a shop
* Open that shop
* Place an order and use the urban piper payment method
* Validate
* Try closing the session
> Observation: Nothing happen, we're now unable to select the menu. If doing
the same steps but in debug mode, a traceback is shown idicating a duplicate
key in a t-foreach

Why the fix:
------------
The duplicate key is happening when rendering the closing popup because the
function `get_closing_control_data()` was returning twice the information related
to the urban piper payment method.

Even though urban piper wasn't build with the use case in mind we want to
keep it possible. We avoid sending twice the urban piper methods if they were
already computed in the super method.

opw-4874699

closes odoo/enterprise#90725

X-original-commit: b2ffe775f79cc8544ca2573cae100747926ff466
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>