[FIX] l10n_br_edi_pos: fix flaky tours
The tours regularly fail in single app builds because the "No other
products found" notification is not found. Waiting on that
notification is necessary because the product search happens
asynchronously.

I was not able to reproduce the test failure locally. But in an
attempt to improve the situation this changes the approach to use the
little "loading" spinner to wait on the search to finish by using
Chrome.isSynced(). This method is used elsewhere in POS tests and
should hopefully be less flaky.

Runbot error 224202

closes odoo/enterprise#89449

X-original-commit: 87e0c29d1e376de7fabe0821e0a141476be9e6cc
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>