[FIX] account_accountant: attachment on reconciled lines
In this commit: https://github.com/odoo/enterprise/commit/13d527b2f940d2215a3bba578e47f180a0b7126e
We wrongly removed the move_attachment_ids from the model params. Meaning that
when putting an attachment on the reconciled line, the paper clip for the
attachments was not visible on the statement line.

Also changing the type of the field to move_attachment_ids to one2many. Since
the field is indeed an One2many. This change transform move_attachment_ids to an
array of id to a static list with a count.

no task id

closes odoo/enterprise#88093

X-original-commit: 8e0d9ff66bd7626db87559b01dbc4971e65b3f91
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>