[FIX] account_accountant: reconcile remainingAmount computation
In this commit: https://github.com/odoo/enterprise/commit/74306d39479f9cfb24bb749a520c533df3f7e404
change the computation of the remaining amount of the reconcile wizard.
Since selectedLineSum compute the sum of the line selected when adding 2 lines
the remaining amount would be changed and kept in the state. Then adding a line
again, the selectedLineSum would be the sum of the 3 lines and not just the new
line that we just added.

no task id

closes odoo/enterprise#88601

X-original-commit: 6d90ef87051fd1aa4a375011681ca1c736c661c9
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>