[IMP] account_reports: analyze tables in `test_balance_sheet_balanced`
This has to be done right after l10n setup, it has no to negative
effect if done above the COA loop or before setting up localisation.

This seems to most affect the domain engine as it can do a lot of
relatively well filtered queries, but those need to be properly
planned. The repeated `ANALYZE` are not free, but trade
favorably (test runs about 4mn faster locally).

closes odoo/enterprise#88207

X-original-commit: 5f753b773f39e6a58059e14d1dc8a26cfd78ffb1
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>