[FIX] website_studio: enable form access on save
When [1] converted the form access option to fit inside `html_builder`,
the form access was supposed to be enabled upon saving a form, but the
used selector was incorrect, which led to the form access not being
enabled.

Also, a `ModelCache` was created which kept track of the same model but
according to various ways to write this key.
This system was flawed and led to reaching distinct object depending
on the used "synonym" key.

This commit:
- fixes the selector used during the save operation
- removes `ModelCache` and uses a plain cache with the model name as key
in all situations
- makes changes to form access included in the history
- adds a tooltip on that confusing option

[1]: https://github.com/odoo/enterprise/commit/689a2186fbe3b932b90bef3b64716c8443bad551

task-4367641

closes odoo/enterprise#89021

X-original-commit: 89fc1023c146a83809f8cfa202177fe44a1aa25a
Signed-off-by: Francois Georis (fge) <fge@odoo.com>
Signed-off-by: Benoit Socias (bso) <bso@odoo.com>