[IMP] account_asset,hr_payroll_account: adjust storno flows
This commit changes entry reversals when storno accounting is enabled,
which required changes in how debit/credit amounts are computed.

To handle asset sale, is_storno flag is set for all lines except for gain/loss line.

To handle reversals of paid payslips in hr payroll, a check is added in the computation of
debit/credit of the move lines. The conditions checks that the company is storno and the move
is a credit note. The check is added in the `hr_payroll_account` module.

Meant by storno accounting: any reversal of a previous entry shall be considered a negative amount
in the original debit/credit position rather than a positive amount in the opposite column.

task-4845062

closes odoo/enterprise#88074

Related: odoo/odoo#214872
Related: odoo/upgrade#8308
Signed-off-by: John Laterre (jol) <jol@odoo.com>