[FIX] pos_enterprise: traceback done order in debug
Steps to reproduce:
- Open a preparation display in debug mode
- Make an order that will appear on the preparation display
- Click on the done button before 5 seconds
- Traceback will appear 5 seconds after the order was received

Issue:
When receiving an order useDelayedValueChange is called in the setup of the
Orderline component. This method will try to access the props after a timeout.
When the preparation orders are deleted in debug mode, the states are deleted
from the front end. Since the props of orderline are the preparation state
linked to it a traceback arises.

Fix:
A check is added to check if the state record exists.

closes odoo/enterprise#89251

X-original-commit: 7defed7503f19b2c1da33cbb4074e3d8a0295bf8
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Loukas Wets (lowe) <lowe@odoo.com>