[REV] hr_contract_salary: unarchive contract on partial signature
This commit reverts the changes made in the following PR: odoo#80136.
This change implied heavy changes in processes, which are unnecessary. The behavior is reverted to unarchive offer contracts upon applicant signature.
An upgrade script is necessary to unarchive all offer contracts that are partially signed (only signed by the applicant).

task-4805608

closes odoo/enterprise#87876

X-original-commit: 286c6dd
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>