[FIX] esg: prevent error when removing the parent emission source
Currently, an error occurs when attempting to remove the parent emission
source from an emission source.

Steps to Reproduce:

 - Install the ESG module.
 - Go to Emission Sources and create a new emission source with a parent
   emission source.
 - Remove the parent emission source.

ValueError: Compute method failed to assign esg.emission.source(<NewId origin=3>,).activity_flow_direct_indirect

This error occurs when the parent emission source is removed, causing the
scope to become False in _compute_scope[1]. Subsequently, in
_compute_activity_flow[2], none of the cases match, which results in a failure
to assign the activity flow and raises an error.

[1] https://github.com/odoo/enterprise/blob/1c7620d8de97757abda1993822f92590987bc46c/esg/models/esg_emission_source.py#L75

[2] https://github.com/odoo/enterprise/blob/1c7620d8de97757abda1993822f92590987bc46c/esg/models/esg_emission_source.py#L77-L78

This commit ensures that if the parent emission source is removed, the
activity flow is properly maintained.

sentry-6690968543

closes odoo/enterprise#88997

X-original-commit: ca37eca1497475aad2bd7400389c756229a0b0de
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>