[FIX] documents_hr: force no owner for hr_employee_folder_id
By default, a document is owned by the user that creates it.
Hr documents folders handle by the system should not be owned by anyone.

In the process of salary configurator, the employee is created while the
current user is the future employee.
(Probably since the split between sudo and with_user)
So the user creates its own employee, and thus, its own employee
folder, and becomes the owner.

In this commit, we make sure that the owner of the employee folder is not set.

Task-4869292

X-original-commit: b28273f7ece7ca4d7cdd83729c06ab29511acd60
Part-of: odoo/enterprise#93960
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>