[FIX] account_reports: returns: don't track return state on checks
Tracking it caused perf issues when resetting the return, because each check needed to run an SQL query to add the tracking information to its chatter. On a clean database, with just the demo data, it took 3 seconds. We're planning to resimplify the UI a little bit in the future, possibly getting rid of the check chatters. In the meantime, we'll just revert this tracking.

closes odoo/enterprise#88170

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>