[FIX] account_loans: error compute with audit trail
Steps:

- Activate Audit Trail in Accounting configuration
- Create and confirm a loan, having some moves posted in the past
- Cancel it, reset it to draft the re-confirm it again
-> ValueError: Expected singleton

This is because posted moves can not be unlinked when canceling/closing
the loan if audit trail is activated. Therefore we try to access the `state`
field from a recordset instead of a singleton.

With this commit, we filter the generated move to get the newly created
one.

opw-4834361

closes odoo/enterprise#90705

X-original-commit: 486d41aa7b66103ed25d38d24eaac1126cf42847
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>
Co-authored-by: Andrea Grazioso (agr-odoo) <agr@odoo.com>