[MOV] account_intrastat: move l10n_be_intrastat generic checks in account_intrastat
In l10n_be_intrastat we were doing some checks such as Vat Exclusive ,
or Only goods included on the intrastat goods which are very generic
checks.

We want those checks to be included in account_intrastat like those for
the EC Sales List since we are going to create Intrastat Returns for a
lot of localizations who will need those checks.

task-4776230

Part-of: odoo/enterprise#85470
Related: odoo/odoo#222598
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>