[FIX] l10n_de_reports: remove null kz83
When generating the German tax report, field Kz83 was included in the XML even when the value was 0.00. This caused the exported file to be rejected by the ELSTER platform.

This fix removes Kz83 from the XML when its value is null or zero. The ELSTER platform automatically calculates this field if it's missing.

source: https://help.deltek.com/product/maconomy/documentation/BPMReporting/CountryReports/Appendix_Germany.html

opw-4783544

closes odoo/enterprise#88332

X-original-commit: 2f67035b4668a805f4107865259a6829a5c6a0ee
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Guillaume Teboul-Tornezy (gute) <gute@odoo.com>