[IMP] l10n_hr_reports: Implement EC Sales List for Croatian Tax returns
Following the implementation of the Tax returns feature in 18.3 which
was focused on the general implementation and Belgium, we want to add
the specification for other country and in this case Croatia.

EC Sales List is a part of this, with VAT Returns as the EC Sales List
is generated when the VAT Return is submitted.

We are showing modals with the instructions as well as a Download PDF
button to make it easier for them to have access to the PDF since it's
important for the users to have the instructions present when
submitting those documents.

This also includes the deadlines and frequencies for the generation of
those documents, and also the translation of the instructions.

task-4776230

Part-of: odoo/enterprise#85470
Related: odoo/odoo#222598
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>