[FIX] planning: avoid altering mutable parameter
Before this commit, the `additionalContext` parameter is altered in
planning gantt renderer to add a context in a specific context which is
not a good idea in general.

This commit makes sure the `additionalContext` is copied before altering
the object to be sure to not have unexpected behavior.

closes odoo/enterprise#85945

Related: odoo/odoo#210758
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>