[IMP] hr_payroll: pay run card standard
Problem
----------
Change the pay run in the kanban view to stay standard
Use a custom component for the pay run card displayed in the filtered list view of payslips.

Objective
----------
Unlink the card in the kanban and the card in the payslip list filtered
So each one can be modified separately.

Solution
----------
A template is defined in the kanban view of payrun. The menu template uses a new compiler to display buttons depending on the screen size in the payrun card.

For the list view of payslip filtered by pay run.
The card is a custom component call in a Record component to fetch the payslip_run record.
Kanban arch is used to fetch fields.
The card and menu template can be used too.
But as we choose to unlink the 2 payrun card, custom component is used with the same PayrunButtonBox Component to display buttons, with all buttons duplicated

task-4873744

closes odoo/enterprise#90135

Related: odoo/upgrade#8061
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>