[IMP] iot: update IoT box server URL on change
Before this commit, if a client renamed their SaaS database to use a
different URL, it would cause paired IoT boxes to fail as they would
try to contact the old URL to send devices etc.

After this commit, when the `web.base.url` parameter is updated (which
will happen the first time an admin logs in from the new URL), we send
a websocket message to all IoT boxes informing them of the new URL.

We don't do this if the new URL is localhost, as this would break the
pairing.

task-4879313

closes odoo/enterprise#88260

Related: odoo/odoo#215407
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>