[FIX] l10n_be_disallowed_expenses: Only write on existing accounts
In the post-init script of `l10n_be_disallowed_expenses` we try to write
the `disallowed_expenses_category_id` field on all the accounts with
xmlids in the data/template/account.account-be.csv file.

This will fail if the account doesn't exist.

Fix: Only write on the accounts that exist. The user can reload the CoA
to create any accounts that don't exist yet if needed.

task-none

closes odoo/enterprise#87598

Related: odoo/odoo#214151
Related: odoo/upgrade#7891
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>