[FIX] l10n_mx_edi_pos: prevent negative residual with eWallet payments
Before this commit, paying an order with eWallet could result in a
small negative residual amount (due to tax calculation), instead
of zero. This prevented users from proceeding to the payment screen.

With this commit, the "lt" function is used to compare amounts while
taking rounding precision into account, ensuring the residual is
treated as zero when appropriate.

opw-4867365

closes odoo/enterprise#92379

X-original-commit: 3bb475d418676f53f15a405b6ecadd52c2fb195b
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>