[FIX] l10n_de_intrastat: missing elements in German Intrastat-Report
In the German Intrastat-Report some elements are missing when downloding the xml report
First, the XML version and character encoding must be included in the XML prolog
Second, the element testIndicator is missing when its false
Third, the element interchangeAgreementId under Party
Fourth, the element regionCode under Declaration > Item

`interchangeAgreementId` field is missing entirely form Odoo, and absence of one of these elemetns prevents
the report from being accepted by German Federal Statistical Office via INSTAT/XML.
To solve this we need to add a new field for interchangeAgreementId in the res_company

steps to reproduce:

Go to Accounting → Configuration → Settings
Under Intrastat, turn on the feature.
Set up a Products for Intrastat:
. Net Weight (per unit)
. under accounting option set Commodity (CN) Code
create an invoice with an EU Partner with the created product and post it.
Go to Reporting → Intrastat Report → Download in XML

opw-4752415

closes odoo/enterprise#88344

Signed-off-by: John Laterre (jol) <jol@odoo.com>