[PERF] planning: import public holidays
The current code of `_recompute_shifts_in_leave_periods` explicitely
calls the compute method `_compute_allocated_hours` of planning.slot
to recompute the allocated hours according to modifications done to the
public holidays. Calling the method directly like that is far from ideal
as we are missing proper batching, fields protection, etc.. that the ORM
usually does for computed fields.

This commit replaces the explicit call by an `add_to_compute`. This
essentially marks the field as "to be recomputed". This recomputation
then happens at the next flush_model/recompute_model call and can
applies the proper ORM optimizations of batching by PREFETCH_MAX size,
protecting fields to avoid triggering loops of recomputations. This
speeds up the import of new Public Holidays in an already populated
database.

speedup

In a customer database with 50K planning.slots, 1300 of which needs to
recompute their allocated_hours, the import timing Before PR vs After PR
is the following:
   40 min -> 4 min

opw-4782452

closes odoo/enterprise#88510

X-original-commit: 990ac1e423d3de538d443e6d257a40c053ed4cc4
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Aurélien van Delft (avd) <avd@odoo.com>