[FIX] planning: allow internal user to access shift calendar view
Before this commit, the internal user without any access to planning app
can no longer access to the calendar view of planning.slot model to see
his shifts and the open shifts. The reason is because the new filter
added is only accessible to planning user.

This commit hides the new filter on resources when the user has no
access to planning app.

Steps to reproduce:
==================
0. install Planning app
1. Create an internal user without any access to planning app (or remove
   planning access to Marc demo)
2. Log as Marc demo (or new user created in step 2)
3. Go to planning app and switch to calendar view

Expected Behavior:
-----------------
The user should be able to access to the calendar view without any
issue.

Current Behavior:
----------------
An access error is raised because the current user cannot access to
`planning.calendar.resource` model which is only accessible to users
with planning access rights.

runbot-error-227622

closes odoo/enterprise#88572

X-original-commit: 3c90dd2bc2f5c753666355dc66f8027d9e08fc1b
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>