[FIX] website_appointment: fix cards params
After the new design applied for the appointments pages, the
propagation of the params was removed when clicking on the
appointments cards.
This was leading to issue to access the corresponding records
when sharing a link to multiple appointments for which the user
didn't have access like the unpublished ones.

Related commit: e94c6fad2d8508942273e8caad2619d149b6fb5d

task-4893791

closes odoo/enterprise#88846

X-original-commit: 91165b073285e8efda71984f1908d0fbe8f5661b
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>