[FIX] sign: validation with sms
To reproduce:
=============
- Activate and set up the SMS authentication for the Sign app.
- Send a document for signature request with sms validation for a contact.
- Go sign the document via the signature request link, sign and validate.
- validate the recieved code and you will get the error:
'Error message: Sorry, an error has occurred. Please complete the document again.'

Problem:
========
when constructing the signature request link, we retrieve the `accessToken`
from `props`, while `accessToken` is not defined in the `props` of the
`Document` component.

Solution:
=========
`accessToken` is defined directly in the `Document` component we can retreive
it from `this.accessToken`

opw-4813339

closes odoo/enterprise#88866

X-original-commit: 03d565b93f26e1a37a2d97d5ba339a34d0d58091
Signed-off-by: Abdelouahab Laaroussi (abla) <abla@odoo.com>