[FIX] helpdesk: replace shared list in _sla_generate_status_values method
Prior to this commit, the `_sla_generate_status_values` method in the `helpdesk.ticket` model was using a shared list for storing SLAs per ticket.
This could lead to unexpected behavior when multiple tickets are processed simultaneously, as they would share the same list instance.
We then removed that logic and replaced it by directly excluding the SLAs of the ticket status having a reached deadline, preventing unnecessary recomputation.

closes odoo/enterprise#88833

X-original-commit: 3fa320ab989c8fdcdc5ed4638ddff809cef6ae5d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Maxime de Neuville (mane) <mane@odoo.com>